+++ js/data.js
@@ -1,0 +1,52 @@
+export const DATA = {
+  postAmount: 25,
+  likes: {
+    min: 15,
+    max: 200
+  },
+  descriptions: [
+    'Почему я такой богатый? Я не свистел дома',
+    'Не насос*ла, а подарили',
+    'Страшно, очень страшно. Если бы знали, что это такое, но мы не знаем, что это такое',
+    'Эй Богдан. - Я здесь.',
+    'Я сегодня потерял друга Саню. Ввел его имя как пароль, а он оказался ненадежный',
+    'Тропикана женщина, горяча и бешена',
+    'В конце концов среди концов нейдешь конец ты наконец',
+    'Если хочешь забыть, забудь',
+    'У каждого волка есть шкура, но не у каждой шкуры есть волк',
+    'Не будь эгоистом, думай о себе',
+    'Сильное заявление. Проверять я его, конечно, не буду',
+    'Необъяснимо, но факт',
+    'Лучше быть последним- первым, чем первым- последним',
+    'Не важно кто, важно кто',
+    'Это вишенка на торте хайпа',
+    'Меня сложно найи, легко потерять и невозможно забыть',
+    'Люблю дождь, так как в нем можно скрыть свои слезы',
+    'Я не конфетка, чтобы быть всем по вкусу. Я орешек, который не всем по зубам',
+    'Сегодня меня никто не поздравил с днем рождения. Ничего. Он у меня не сегодня',
+    'Зима близко',
+    'Не можешь найти работу? Умножь мощность на время',
+    'Почему молодому разработчику не позволено вращать коесико мыши, а разработчику постарше - позволено? Потому что Elder Scrolls',
+    'Почему все хакеры заказывают одну и ту же проститутку. Они пытаются ее задудосить',
+    'От чего не избавит спортзал айтишника? от JIRA',
+    'Если хочешь идти, иди.',
+  ],
+  comments: [
+    'Это рофл',
+    'Кринж на кринже и кринжем погоняет',
+    'Лайк, подписка',
+    'Трэш',
+    'Хочешь скину ссылку?',
+    'Пуньк захуюньк',
+    'На чиле, на расслабоне',
+  ],
+  names: [
+    'Кеша',
+    'Глаша',
+    'Клава',
+    'Стрегобор',
+    'Петя',
+    'Иммануил',
+    'Дибон',
+  ]
+};
