+++ js/main.js
@@ -1,95 +1,1 @@
-//Функция, возвращающая случайное целое число из переданного диапазона включительно.
-const getRandInt = (min, max) => {
-  if (min < 0 || max < 0) {
-    return 'Рип. Оба числа должны быть больше нуля';
-  }
-  return Math.floor(min + Math.random() * (max - min + 1));
-};
-
-//Функция для проверки максимальной длины строки. Будет использоваться для проверки длины введённого комментария, но должна быть универсальна.
-const getStrLen = (str, maxLen) => str.length <= maxLen;
-getRandInt(3, 4);
-getStrLen('ghjfdk dfjjhg jfgh', 4);
-
-//напишите необходимые функции для создания массива из 25 сгенерированных объектов
-const DATA = {
-  descriptions: [
-    'Почему я такой богатый? Я не свистел дома',
-    'Не насос*ла, а подарили',
-    'Страшно, очень страшно. Если бы знали, что это такое, но мы не знаем, что это такое',
-    'Эй Богдан. - Я здесь.',
-    'Я сегодня потерял друга Саню. Ввел его имя как пароль, а он оказался ненадежный',
-    'Тропикана женщина, горяча и бешена',
-    'В конце концов среди концов нейдешь конец ты наконец',
-    'Если хочешь забыть, забудь',
-    'У каждого волка есть шкура, но не у каждой шкуры есть волк',
-    'Не будь эгоистом, думай о себе',
-    'Сильное заявление. Проверять я его, конечно, не буду',
-    'Необъяснимо, но факт',
-    'Лучше быть последним- первым, чем первым- последним',
-    'Не важно кто, важно кто',
-    'Это вишенка на торте хайпа',
-    'Меня сложно найи, легко потерять и невозможно забыть',
-    'Люблю дождь, так как в нем можно скрыть свои слезы',
-    'Я не конфетка, чтобы быть всем по вкусу. Я орешек, который не всем по зубам',
-    'Сегодня меня никто не поздравил с днем рождения. Ничего. Он у меня не сегодня',
-    'Зима близко',
-    'Не можешь найти работу? Умножь мощность на время',
-    'Почему молодому разработчику не позволено вращать коесико мыши, а разработчику постарше - позволено? Потому что Elder Scrolls',
-    'Почему все хакеры заказывают одну и ту же проститутку. Они пытаются ее задудосить',
-    'От чего не избавит спортзал айтишника? от JIRA',
-    'Если хочешь идти, иди.',
-  ],
-  comments: [
-    'Это рофл',
-    'Кринж на кринже и кринжем погоняет',
-    'Лайк, подписка',
-    'Трэш',
-    'Хочешь скину ссылку?',
-    'Пуньк захуюньк',
-    'На чиле, на расслабоне',
-  ],
-  names: [
-    'Кеша',
-    'Глаша',
-    'Клава',
-    'Стрегобор',
-    'Петя',
-    'Иммануил',
-    'Дибон',
-  ]
-};
-
-const getComments = () => {
-  const commentsList = [...DATA.comments];
-  const commentsTotal = getRandInt(1, DATA.comments.length);
-  const commentsPost = [];
-  while (commentsPost.length < commentsTotal) {
-    const getRandomComment = getRandInt(0, commentsList.length - 1);
-    commentsPost.push({
-      id: commentsPost.length,
-      name: DATA.names[commentsPost.length],
-      avatar: `photos/${commentsPost.length}.svg`,
-      comment: commentsList[getRandomComment],
-    });
-    commentsList.splice(getRandomComment, 1);
-  }
-  return commentsPost;
-};
-
-const createPosts = () => {
-  const posts = [];
-  for (let i = 1; i <= 25; i++) {
-    const post = {
-      id: i,
-      url: `photos/${i}.jpg`,
-      description: DATA.descriptions[i],
-      likes: getRandInt(15, 200),
-      comments: getComments(),
-    };
-    post.push(post);
-  }
-  return posts;
-};
-
-createPosts();
+import "./posts.js";
